use float64 for int pow precision, enable test_pow_int

+++ tinygrad/tensor.py
@@ -3465,8 +3465,8 @@ class Tensor(OpMixin):
     base, exponent = self._broadcasted(x, reverse=reverse)
     # NOTE: negative exponents not checked (would require sync)
     if not base.is_floating_point() and isinstance(x, Tensor) and not exponent.is_floating_point():
-      float_base = base.cast(dtypes.float32)
-      float_exp = exponent.cast(dtypes.float32)
+      float_base = base.cast(dtypes.float64)
+      float_exp = exponent.cast(dtypes.float64)
       ret = float_base._apply_uop(UOp.pow, float_exp)
       return ret.round().cast(base.dtype)
     if not base.is_floating_point() and not (isinstance(x, int) and x >= 0): raise RuntimeError("base needs to be float")
